Clean GEMMSCanvas

+++ GEMMS/src/main/java/ch/heigvd/layer/GEMMSCanvas.java
@@ -2,25 +2,18 @@ package ch.heigvd.layer;
 
 import ch.heigvd.gemms.CSSIcons;
 import ch.heigvd.workspace.LayerListable;
-
 import java.io.IOException;
-
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
-
 import javafx.geometry.Point3D;
 import javafx.scene.SnapshotParameters;
-
 import javafx.scene.canvas.GraphicsContext;
-import javafx.scene.image.Image;
 import javafx.scene.image.PixelReader;
 import javafx.scene.image.PixelWriter;
 import javafx.scene.image.WritableImage;
 import javafx.scene.paint.Color;
-import javafx.scene.transform.MatrixType;
 import javafx.scene.transform.Rotate;
 import javafx.scene.transform.Transform;
-import javafx.scene.transform.Translate;
 
 public class GEMMSCanvas extends javafx.scene.canvas.Canvas implements IGEMMSNode, LayerListable {
 
@@ -45,22 +38,19 @@ public class GEMMSCanvas extends javafx.scene.canvas.Canvas implements IGEMMSNod
         s.defaultWriteObject();
 
         // Get the size
-        int width = (int) getWidth();
-        int height = (int) getHeight();
+        double width = getWidth();
+        double height = getHeight();
 
         // Write the size
-        s.writeInt(width);
-        s.writeInt(height);
+        s.writeDouble(width);
+        s.writeDouble(height);
 
         // Get an image 
         SnapshotParameters sp = new SnapshotParameters();
         sp.setFill(Color.TRANSPARENT);
-
-
         sp.setTransform(getLocalToSceneTransform());
 
-
-        WritableImage writableImage = new WritableImage(width, height);
+        WritableImage writableImage = new WritableImage((int)width, (int)height);
         snapshot(sp, writableImage);
 
         // Get a pixel reader
@@ -118,8 +108,8 @@ public class GEMMSCanvas extends javafx.scene.canvas.Canvas implements IGEMMSNod
     private void readObject(ObjectInputStream s) throws IOException, ClassNotFoundException {
 
         // Get the size of the canvas
-        int width = s.readInt();
-        int height = s.readInt();
+        double width = s.readDouble();
+        double height = s.readDouble();
 
         // Set the size of this canvas
         setWidth(width);
